Moved pmd loading code to base library.
